Make plain node package
Strip out basically everything, and assume nothing.
