--- config/environment.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports = function(/* environment, appConfig */) {
-  return { };
-};
