+++ package.json
@@ -3,70 +3,21 @@
   "version": "0.0.3",
   "description": "Sinon assertions for QUnit with custom messages support.",
   "keywords": [
-    "ember-addon"
+    "qunit",
+    "sinon"
   ],
   "repository": "https://github.com/PrecisionNutrition/qunit-sinon-assertions",
   "license": "MIT",
   "author": "Precision Nutrition Team",
   "directories": {
+    "lib": "lib",
     "doc": "doc",
     "test": "tests"
   },
-  "scripts": {
-    "build": "ember build --environment=production",
-    "lint": "npm-run-all --aggregate-output --continue-on-error --parallel lint:*",
-    "lint:hbs": "ember-template-lint .",
-    "lint:js": "eslint .",
-    "start": "ember serve",
-    "test": "npm-run-all lint:* test:*",
-    "test:ember": "ember test",
-    "test:ember-compatibility": "ember try:each"
-  },
-  "dependencies": {
-    "broccoli-funnel": "^3.0.3",
-    "ember-cli-babel": "^7.19.0",
-    "ember-cli-htmlbars": "^4.3.1"
-  },
-  "devDependencies": {
-    "@ember/optional-features": "^1.3.0",
-    "@glimmer/component": "^1.0.0",
-    "@glimmer/tracking": "^1.0.0",
-    "babel-eslint": "^10.1.0",
-    "broccoli-asset-rev": "^3.0.0",
-    "ember-cli": "~3.18.0",
-    "ember-cli-dependency-checker": "^3.2.0",
-    "ember-cli-inject-live-reload": "^2.0.2",
-    "ember-cli-sri": "^2.1.1",
-    "ember-cli-uglify": "^3.0.0",
-    "ember-disable-prototype-extensions": "^1.1.3",
-    "ember-export-application-global": "^2.0.1",
-    "ember-load-initializers": "^2.1.1",
-    "ember-maybe-import-regenerator": "^0.1.6",
-    "ember-qunit": "^4.6.0",
-    "ember-resolver": "^8.0.0",
-    "ember-source": "~3.18.0",
-    "ember-source-channel-url": "^2.0.1",
-    "ember-template-lint": "^2.6.0",
-    "ember-try": "^1.4.0",
-    "ember-welcome-page": "^4.0.0",
-    "eslint": "^6.8.0",
-    "eslint-plugin-ember": "^8.4.0",
-    "eslint-plugin-node": "^11.1.0",
-    "loader.js": "^4.7.0",
-    "npm-run-all": "^4.1.5",
-    "qunit-dom": "^1.2.0"
-  },
+  "scripts": {},
+  "dependencies": {},
+  "devDependencies": {},
   "engines": {
     "node": "10.* || >= 12"
-  },
-  "ember": {
-    "edition": "octane"
-  },
-  "ember-addon": {
-    "configPath": "tests/dummy/config",
-    "after": [
-      "ember-cli-qunit",
-      "ember-qunit"
-    ]
   }
 }
