--- tests/dummy/app/router.js
@@ -1,10 +1,0 @@
-import EmberRouter from '@ember/routing/router';
-import config from './config/environment';
-
-export default class Router extends EmberRouter {
-  location = config.locationType;
-  rootURL = config.rootURL;
-}
-
-Router.map(function() {
-});
