--- tests/dummy/app/templates/application.hbs
@@ -1,5 +1,0 @@
-{{!-- The following component displays Ember's default welcome message. --}}
-<WelcomePage />
-{{!-- Feel free to remove this! --}}
-
-{{outlet}}
