--- tests/test-helper.js
@@ -1,8 +1,0 @@
-import Application from '../app';
-import config from '../config/environment';
-import { setApplication } from '@ember/test-helpers';
-import { start } from 'ember-qunit';
-
-setApplication(Application.create(config.APP));
-
-start();
